+++ backend/garpix_notify/utils/__init__.py
@@ -1,5 +1,10 @@
 from .file import get_file_path
-from .sms_checker import SMSClient
-from .receiving import receiving_users
-from .call_code_cheker import CallClient
+from .receiving import ReceivingUsers
 from .send_data import notify_create
+
+from .call_client import CallClient
+from .sms_client import SMSClient
+from .email_client import EmailClient
+from .telegram_client import TelegramClient
+from .viber_client import ViberClient
+from .push_client import PushClient
