+++ backend/garpix_notify/utils/receiving.py
@@ -2,52 +2,65 @@ from django.contrib.auth import get_user_model
 from django.db.models import Q
 
 
-def receiving_users(users_list, value=False):
-    User = get_user_model()
-    for user_list in users_list:
+class ReceivingUsers:
+
+    def __init__(self, users_list, value=None):
+        self.users_list = users_list
+        self.value = value
+
+    def __returning_specific_list(self, receivers):
+        # Убираем дубликаты пользователей для переданного значения из функции
+        receivers_dict = list({v[self.value]: v for v in receivers}.values())
+        receivers_new = [user[self.value] for user in receivers_dict]
+        return receivers_new
+
+    def __receiving_users(self):
+        user_model = get_user_model()
         receivers = []
-        # Сначала проверям есть ли дополнительные списки получателей
-        for participant in user_list.participants.all():
-            if participant.user or participant.email:
-                receivers.append({
-                    'user': participant.user,
-                    'email': participant.user.email if participant.user else participant.email,
-                    'phone': participant.user.phone if participant.user else participant.email,
-                    'viber_chat_id': participant.user.viber_chat_id if participant.user else participant.viber_chat_id
-                })
-        # Проверяем рассылку, не отмечена ли массовая рассылка
-        if user_list.mail_to_all is False:
-            group_users = User.objects.filter(
+        for user_list in self.users_list:
+            # Проверяем рассылку, не отмечена ли массовая рассылка
+            if user_list.mail_to_all:
+                users = user_model.objects.all()
+                receivers.append(
+                    {
+                        'user': user,
+                        'email': user.email,
+                        'phone': user.phone,
+                        'viber_chat_id': user.viber_chat_id,
+                    } for user in users)
+
+                if self.value:
+                    return self.__returning_specific_list(receivers)
+                return receivers
+            # Собираем данные из дополнительного списка получателей, если он есть.
+            users_participants = user_list.participants.all()
+            if users_participants.exists():
+                for participant in users_participants:
+                    if participant.user or participant.email or participant.phone or participant.viber_chat_id:
+                        receivers.append({
+                            'user': participant.user,
+                            'email': participant.user.email if participant.user else participant.email,
+                            'phone': participant.user.phone if participant.user else participant.phone,
+                            'viber_chat_id':
+                                participant.user.viber_chat_id if participant.user else participant.viber_chat_id
+                        })
+            group_users = user_model.objects.filter(
                 Q(groups__in=list(
                     user_list.user_groups.all()
                 )))
-            for user in group_users:
-                receivers.append(
-                    {
-                        'user': user,
-                        'email': user.email,
-                        'phone': user.phone,
-                        'viber_chat_id': user.viber_chat_id,
-                    }
-                )
-        else:
-            users = User.objects.all()
-            for user in users:
-                receivers.append(
-                    {
-                        'user': user,
-                        'email': user.email,
-                        'phone': user.phone,
-                        'viber_chat_id': user.viber_chat_id,
-                    }
-                )
-        # Убираем дубликаты пользователей для переданного значения из функции
-        if value:
-            receivers = list({v[value]: v for v in receivers}.values())
-            receivers_new = []
-            for user in receivers:
-                if user[value]:
-                    receivers_new.append(user[value])
-            return receivers_new
 
-        return receivers
+            receivers.append(
+                {'user': user,
+                 'email': user.email,
+                 'phone': user.phone,
+                 'viber_chat_id': user.viber_chat_id
+                 } for user in group_users)
+
+            if self.value:
+                return self.__returning_specific_list(receivers)
+
+            return receivers
+
+    @classmethod
+    def run_receiving_users(cls, users_list, value=None):
+        return cls(users_list, value).__receiving_users()
